patient_service/requirements.txt
Django
mysqlclient>=2.2.0
djangorestframework
djangorestframework-simplejwt
requests
gunicorn
